Revert "Removed Redis for user status from backend"
This reverts commit edd75b35feb0dbfdbfdc0e3e68c7c3ed2cd0e384.

+++ backend/src/main/java/hr/tvz/ljubojevic/chatterbox/controller/AuthController.java
@@ -4,12 +4,18 @@ import hr.tvz.ljubojevic.chatterbox.DTO.AuthRequestDTO;
 import hr.tvz.ljubojevic.chatterbox.DTO.JwtResponseDTO;
 import hr.tvz.ljubojevic.chatterbox.DTO.RefreshTokenDTO;
 import hr.tvz.ljubojevic.chatterbox.model.RefreshToken;
+import hr.tvz.ljubojevic.chatterbox.model.User;
+import hr.tvz.ljubojevic.chatterbox.repository.UserRepository;
 import hr.tvz.ljubojevic.chatterbox.service.auth.JwtService;
 import hr.tvz.ljubojevic.chatterbox.service.auth.RefreshTokenService;
+import hr.tvz.ljubojevic.chatterbox.service.user.UserStatusService;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,10 +28,25 @@ public class AuthController {
     private JwtService jwtService;
     private RefreshTokenService refreshTokenService;
 
+    @Autowired
+    private UserStatusService userStatusService;
+
+    @Autowired
+    private WebSocketMessageController webSocketController;
+
+    @Autowired
+    private UserRepository userRepository;
+
     @PostMapping("/api/login")
     public JwtResponseDTO authenticationAndGetToken(@RequestBody AuthRequestDTO authRequestDTO) {
         Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequestDTO.getUsername(), authRequestDTO.getPassword()));
         System.out.println("login called");
+
+        User user = userRepository.findByUsername(authRequestDTO.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+        userStatusService.setUserStatus(user.getId(), "online");
+        webSocketController.sendStatusUpdate(user.getId(), "online");
 
         if (authentication.isAuthenticated()) {
             RefreshToken refreshToken = refreshTokenService.createRefreshToken(authRequestDTO.getUsername());
@@ -55,7 +76,21 @@ public class AuthController {
 
     @PostMapping("/api/logout")
     public void logout(@RequestBody RefreshTokenDTO refreshTokenDTO) {
-        refreshTokenService.deleteByToken(refreshTokenDTO.getToken());
+        System.out.println("Logout called - " + refreshTokenDTO.getToken());
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if (authentication != null && authentication.getPrincipal() instanceof UserDetails) {
+            UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+            String username = userDetails.getUsername();
+
+            User user = userRepository.findByUsername(username)
+                    .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+            userStatusService.setUserStatus(user.getId(), "offline");
+            webSocketController.sendStatusUpdate(user.getId(), "offline");
+
+            refreshTokenService.deleteByToken(refreshTokenDTO.getToken());
+        }
 
         System.out.println("Logout called and user status updated");
     }
